new file:   src/components/navbar-about.vue 	modified:   src/components/navbar.vue 	modified:   src/views/AboutView.vue 	modified:   src/views/HomeView.vue
